move impl to main class
