+++ Sources/Honeycomb/Honeycomb.swift
@@ -205,6 +205,58 @@ public class Honeycomb {
             MXMetricManager.shared.add(self.metricKitSubscriber)
         }
     }
+    
+    private static let errorLoggerInstrumentationName = "@honeycombio/instrumentation-error-logger"
+
+    private static let defaultErrorLogger = OpenTelemetry.instance.loggerProvider.get(
+        instrumentationScopeName: errorLoggerInstrumentationName
+    )
+
+    protocol AttributeValueConvertable {
+        func attributeValue() -> AttributeValue
+    }
+    
+    static func logError(
+        error: Error,
+        attributes: [String: AttributeValue] = [:],
+        thread: Thread?,
+        logger: OpenTelemetryApi.Logger = defaultErrorLogger
+    ) {
+        let timestamp = Date()
+        let type = String( describing: Mirror(reflecting: error).subjectType)
+        let message = error.localizedDescription
+        
+        var errorAttributes = [
+            "exception.type": type.attributeValue(),
+            "exception.message": message.attributeValue(),
+        ].merging(attributes, uniquingKeysWith: {(_, last) in last})
+        
+        if let name = thread?.name {
+            errorAttributes["exception.thread"] = name.attributeValue()
+        }
+        
+        logError("", errorAttributes, logger, timestamp)
+    }
+
+    private static func logError(
+        _ namespace: String,
+        _ attributes: [String: AttributeValue],
+        _ logger: OpenTelemetryApi.Logger = defaultErrorLogger,
+        _ timestamp: Date = Date()
+    ) {
+        var logAttrs: [String: AttributeValue] = [
+            "name": namespace.attributeValue()
+        ]
+        for (key, value) in attributes {
+            let namespacedKey = "\(namespace).\(key)"
+            logAttrs[namespacedKey] = value
+        }
+        
+        logger.logRecordBuilder()
+            .setTimestamp(timestamp)
+            .setAttributes(logAttrs)
+            .emit()
+    }
 
     @available(iOS 16.0, macOS 12.0, *)
     public static func setCurrentScreen(path: NavigationPath) {
